+++ app/src/main/java/com/deslomator/tagtimer/viewmodel/EventTrashViewModel.kt
@@ -22,7 +22,7 @@ class EventTrashViewModel @Inject constructor(
     private val appDao: AppDao, savedStateHandle: SavedStateHandle
 ): ViewModel() {
 
-    private val _sessionId = MutableStateFlow(0)
+    private val _sessionId = MutableStateFlow(0L)
     private val _state = MutableStateFlow(EventTrashState())
 
     @OptIn(ExperimentalCoroutinesApi::class)
@@ -66,7 +66,7 @@ class EventTrashViewModel @Inject constructor(
     }
 
     init {
-        val sessionId = savedStateHandle.get<Int>("sessionId") ?: 0
+        val sessionId = savedStateHandle.get<Long>("sessionId") ?: 0
         _sessionId.update { sessionId }
         viewModelScope.launch {
             _state.update {
